Remove our own threading stuff.
Back when it was first written, I realized that bugzilla2fedmsg was
going to need *more time* than most consumers to do what it needed to
do.  It needed to receive a message, then make a costly xmlrpc call to
bugzilla itself to get all kinds of details about the message (history,
etc), then it would pack all that stuff inside the message and forward
it along to the fedora infrastructure bus.

To accomplish this, I wrote some threading stuff into bugzilla2fedmsg.
A master thread would receive messages and put them in a queue.  Workers
(each a thread) would take messages off the queue, do their thing, and
then return to wait on the queue.

It worked so well that it was added to fedmsg/moksha core.  Now FMN,
fedbadges, and all their friends benefit from this setup.

The time has come to remove bugzilla2fedmsg's own threading machinery.
As it stands now, fedmsg itself spins up a couple threads for the
bugzilla2fedmsg consumer.  And then each of those
consumer-worker-threads spins up a couple of their *own* worker threads!
There's no need for that second layer.  This commit removes that.