Auto add new packages.
This should be the last missing piece we need before we can deploy
the-new-hotness.

It watches for new packages being added to Fedora (after they've gone
through review, when they get added to pkgdb), and then it checks to see
if any projects are registered in anitya with the same upstream_url.

If none exist, then it adds the new project with the best guess it can
at the backend and it maps the project to the new Fedora package.