Only followup on rawhide builds.
And we define what a rawhide build is in the config.

This will be, like so many other things, a PITA in that when mass branch
time comes along, we'll have to remember to go into the hotness config
and update the suffix to the new rawhide suffix.  Such is life (if we
had a releng microservice that made available what the rawhide tag and
suffix was.. that would be awesome.  We should write that).

This fixes #14.