Handle multiply mapped anitya projects.
It is possible for an anitya project to be mapped to two or more Fedora
packages.  In these cases we will now check the monitoring flag of *each*
one of those packages (not just whichever one appears first in the
list) and then file bugs for each of those packages that **do** want
bugs to be filed.

Fixes #33.