Handle newly-mapped packages from anitya.
@tyll requested this in #34:  sometimes, a project exists on
release-monitoring.org *and* the package exists in Fedora.. but there is
no mapping between the two.  When a user establishes that mapping
between them, the-new-hotness should respond by filing a bug if
necessary.

If the project in question has **no** version information at the time
this message arrives, the we force a new check in anitya and bail..
waiting on the next message from that to arrive.

Fixes #34.