Handle pkgdb.package.update messages.
This should fix #46.

Sometimes, a new package added to Fedora pkgdb doesn't have an
upstream_url off the bat.  We already guard against that and avoid
trying to update anitya with that empty information.

This new code responds when the pkgdb weekly cronjob goes through and
updates the upstream_url field from a scan of the .spec/srpm
information.

Since packagers have indicated that a package has a new upstream url,
we use that to update anitya's homepage information, and potentially map
packages to projects for the first time.