Silence this error email.
And replace it with a warning in the logs.

I looked into all the error emails we got about this the last week, and
they're all reasonable situations.

- A new package was added to pkgdb after the Review process completed.
- hotness noticed that, and added the pkg to anitya.
- hotness forced an anitya check
- anitya published a fedmsg message that a new version was found
- hotness received that message and tried to compare that new upstream
  version against the version in rawhide but... of course, no build had
  completed in rawhide yet.  And even if it had in the tiny interval of
  time that has elapsed so far, the rawhide repo would not have been
  rebuilt for us to check.

So, we should assume that rawhide *will* be up to date with the latest
upstream at this point, since the package **just** came out of package
review.