Fix release-monitoring.org html scraping.
Kinda ugly here, but login with release-monitoring.org broke when we
added a search form to every page there.

Hotness would try to scrape each page for a form, and then extract the
csrf_token from that form.  Except now, there are **two** forms on every
page, so that scraping broke.

This code updates each attempt at that to look explicitly for the form
it is interested in.