Rewrite topics to handle development mode.
This should solve the problems that @phracek hit trying to test
the-new-hotness locally.

The problem is that:

- the-new-hotness has a list of topics it listens to, hardcoded to prod.
- when you replay a message with ``fedmsg-dg-replay`` it constructs a
  new topic for the replayed message, using your local settings (dev
  settings).
- even when you switch your local environment settings from dev to prod,
  the anitya topics have a different *prefix* -- org.release-monitoring
  instead of org.fedoraproject.

This gets around all that by (at startup time) checking to see if you're
in a dev environment, and if so, it rewrites the topics that
the-new-hotness is *listening* for and adds some appropriate ones so
that fedmsg-dg-replay will work.  This shouldn't affect our prod or stg
setups at all.