Make this log statement less verbose.
Any calls to ``log.error`` or ``log.exception`` will generate an email
to the admins of the-new-hotness.  Those are very useful, but if we get
them for non-urgent situations, it can inadvertently train us to not pay
attention to the emails over time.

So, I'd like to propose that we *demote* this log statement to a
warning.  What do you think @phracek?  Should it be kept as an error?