Correct another instance of mis-used six.iteritems().
I'd like @phracek to review this one before it gets merged since it
depends on the structure of the output from rebase-helper.  We found an
erroneous instance like this in #106 and I'm not sure if this is the
same kind of situation or not.

See https://github.com/fedora-infra/the-new-hotness/pull/106#discussion_r55363099