Remove rebase-helper
This removes the dependency on rebase-helper in the-new-hotness. It also
fixes several of the commands that the-new-hotness shells out to which
depended on git to have a user configured. Finally, it also gets rid of
the 'userstring' setting in favor of a 'email_user' tuple which makes
configuring git with an author simpler.

Fixes #145

Signed-off-by: Jeremy Cline <jeremy@jcline.org>