Add tests for helpers.secure_download
This adds pycurl as an explicit requirement. pycurl is now required to be available in the tox environment, which in turn requires both libcurl-devel and openssl-devel installed in the vagrant environment. pycurl is told to use the correct ssl library by setting the PYCURL_SSL_LIBRARY in tox.ini.

The StringIO import has been replaced with io.StringIO, the former no longer exists in Python3.