Save koji id even if build use case fails
Currently the-new-hotness didn't saved the koji build id when any
failure happened during the build use case. This caused that when the
build fedora message arrived it was not recognized by the-new-hotness.

This commit is saving the koji build id to cache even when there is
failure during the build use case, but the build successfully started.

Fixes #332

Signed-off-by: Michal Konečný <mkonecny@redhat.com>